LPS-21191 Added check to prevent NullPointerException

+++ portal-service/src/com/liferay/portal/kernel/io/ByteArrayFileInputStream.java
@@ -66,7 +66,7 @@ public class ByteArrayFileInputStream extends InputStream {
 			data = null;
 			fileInputStream = null;
 
-			if (deleteOnClose) {
+			if (deleteOnClose && file != null) {
 				file.delete();
 			}
